Refactor safety raster workflow WIP #479
