+++ geest/core/algorithms/__init__.py
@@ -5,7 +5,4 @@ from .polygon_per_cell_processor import PolygonPerCellProcessor
 from .single_point_buffer_processor import SinglePointBufferProcessor
 from .safety_polygon_processor import SafetyPerCellProcessor
 from .raster_reclassification_processor import RasterReclassificationProcessor
-from .safety_raster_reclassification_processor import (
-    SafetyRasterReclassificationProcessor,
-)
 from .ors_multibuffer_processor import ORSMultiBufferProcessor
